% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{contDist}
\alias{contDist}
\alias{dcbinom}
\alias{pcbinom}
\alias{qcbinom}
\alias{rcbinom}
\alias{dcpoisson}
\alias{pcpoisson}
\alias{qcpoisson}
\alias{rcpoisson}
\title{Continuous Binomial and Poisson Distributions}
\description{
Functions to perform operations on continuous binomial and Poisson distributions.
}
\details{
Continuous Binomial and Poisson Distributions

These functions provide information about the continuous binomial and Poisson distribution functions,
including density, distribution, quantile functions and random generation.

The continuous binomial distribution is a probability distribution that generalizes the binomial distribution to non-integer values.
The continuous Poisson distribution is a probability distribution that generalizes the Poisson distribution to non-integer values.
}
\examples{
\dontrun{
# Density of continuous binomial distribution
dcbinom(0.5, 1, 0.5)
# Density of continuous Poisson distribution
dcpoisson(0.5, 1)
}
}
\references{
Ilienko, A. (2013): Continuous counterparts of Poisson and Binomial distributions and their properties. Annales Univ. Sci. Budapest., Sect. Comp., 39, 137-147. http://arxiv.org/abs/1303.5990
}
\seealso{
\code{\link[cubature]{adaptIntegrate}}, \code{\link[zipfR]{Rbeta}}, \code{\link[zipfR]{Rgamma}}
}
\author{
Your Name [aut, cre]
}
\keyword{distributions}
