% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcbinom.R
\name{dcbinom}
\alias{dcbinom}
\title{Continuous Binomial Density Function}
\usage{
dcbinom(x, nu, pp)
}
\arguments{
\item{x}{vector of quantiles.}

\item{nu}{number of trials.}

\item{pp}{probability of success on each trial.}
}
\value{
The density of the continuous binomial distribution at each point specified by x.
}
\description{
Computes the density of the continuous binomial distribution.
}
\examples{
\dontrun{
# Density of continuous binomial distribution
dcbinom(0.5, 1, 0.5)
}
}
