% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpoisson.R
\name{dcpoisson}
\alias{dcpoisson}
\title{Continuous Poisson Density Function}
\usage{
dcpoisson(x, lambda)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda}{rate parameter.}
}
\value{
The density of the continuous Poisson distribution at each point specified by x.
}
\description{
Computes the density of the continuous Poisson distribution.
}
\examples{
\dontrun{
# Density of continuous Poisson distribution
dcpoisson(0.5, 1)
}
}
