% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcbinom.R
\name{pcbinom}
\alias{pcbinom}
\title{Continuous Binomial Probability Function}
\usage{
pcbinom(q, nu, pp)
}
\arguments{
\item{q}{vector of quantiles.}

\item{nu}{number of trials.}

\item{pp}{probability of success on each trial.}
}
\value{
The probability of the continuous binomial distribution at each point specified by q.
}
\description{
Computes the probability of the continuous binomial distribution.
}
\examples{
\dontrun{
# Probability of continuous binomial distribution
pcbinom(0.5, 1, 0.5)
}
}
