% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpoisson.R
\name{pcpoisson}
\alias{pcpoisson}
\title{Continuous Poisson Probability Function}
\usage{
pcpoisson(q, lambda)
}
\arguments{
\item{q}{vector of quantiles.}

\item{lambda}{rate parameter.}
}
\value{
The probability of the continuous Poisson distribution at each point specified by q.
}
\description{
Computes the probability of the continuous Poisson distribution.
}
\examples{
\dontrun{
# Probability of continuous Poisson distribution
pcpoisson(0.5, 1)
}
}
