% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcpoisson.R
\name{qcpoisson}
\alias{qcpoisson}
\title{Continuous Poisson Quantile Function}
\usage{
qcpoisson(p, lambda)
}
\arguments{
\item{p}{vector of probabilities.}

\item{lambda}{rate parameter.}
}
\value{
The quantile of the continuous Poisson distribution at each point specified by p.
}
\description{
Computes the quantile of the continuous Poisson distribution.
}
\examples{
\dontrun{
# Quantile of continuous Poisson distribution
qcpoisson(0.5, 1)
}
}
