% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbinom.R
\name{rcbinom}
\alias{rcbinom}
\title{Continuous Binomial Random Generation Function}
\usage{
rcbinom(size, nu, pp)
}
\arguments{
\item{size}{number of observations.}

\item{nu}{number of trials.}

\item{pp}{probability of success on each trial.}
}
\value{
A vector of random deviates from the continuous binomial distribution.
}
\description{
Generates random deviates from a continuous binomial distribution.
}
\examples{
\dontrun{
# Generate random deviates from a continuous binomial distribution
rcbinom(10, 1, 0.5)
}
}
