% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpoisson.R
\name{rcpoisson}
\alias{rcpoisson}
\title{Continuous Poisson Random Generation Function}
\usage{
rcpoisson(size, lambda)
}
\arguments{
\item{size}{number of observations.}

\item{lambda}{rate parameter.}
}
\value{
A vector of random deviates from the continuous Poisson distribution.
}
\description{
Generates random deviates from a continuous Poisson distribution.
}
\examples{
\dontrun{
# Generate random deviates from a continuous Poisson distribution
rcpoisson(10, 1)
}
}
